refactor(file-io): unify atomicWrite with Windows retry
Session manager previously had a private atomicWrite with EPERM/EACCES
retry-on-rename for Windows file-handle contention. The exported
atomicWrite in file-io/index.ts had neither retry nor unlink-on-failure.
Unify on the retry-capable version: extract renameWithRetry, reuse it
from both atomicWrite and atomicWriteBuffer, and have session/manager
import the shared helper. Removes ~20 lines of duplication and gives
docx/markdown writes the same Windows robustness.
